requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-storage
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
plotly
requests
python-dotenv
transformers
torch
pillow
opencv-python
nltk
spacy
wordcloud
textstat
langdetect
networkx
pandas-gbq
db-dtypes